+++ src/Plugin/BufferedAdd/BufferedAdd.php
@@ -234,14 +234,14 @@ class BufferedAdd extends AbstractPlugin
         $commitWithin = null === $commitWithin ? $this->getCommitWithin() : $commitWithin;
 
         $event = new PreFlushEvent($this->buffer, $overwrite, $commitWithin);
-        $this->client->getEventDispatcher()->dispatch(Events::PRE_FLUSH, $event);
+        $this->client->getEventDispatcher()->dispatch($event, Events::PRE_FLUSH);
 
         $this->updateQuery->addDocuments($event->getBuffer(), $event->getOverwrite(), $event->getCommitWithin());
         $result = $this->client->update($this->updateQuery, $this->getEndpoint());
         $this->clear();
 
         $event = new PostFlushEvent($result);
-        $this->client->getEventDispatcher()->dispatch(Events::POST_FLUSH, $event);
+        $this->client->getEventDispatcher()->dispatch($event, Events::POST_FLUSH);
 
         return $result;
     }
@@ -261,7 +261,7 @@ class BufferedAdd extends AbstractPlugin
     public function commit(?bool $overwrite = null, ?bool $softCommit = null, ?bool $waitSearcher = null, ?bool $expungeDeletes = null)
     {
         $event = new PreCommitEvent($this->buffer, $overwrite, $softCommit, $waitSearcher, $expungeDeletes);
-        $this->client->getEventDispatcher()->dispatch(Events::PRE_COMMIT, $event);
+        $this->client->getEventDispatcher()->dispatch($event, Events::PRE_COMMIT);
 
         $this->updateQuery->addDocuments($this->buffer, $event->getOverwrite());
         $this->updateQuery->addCommit($event->getSoftCommit(), $event->getWaitSearcher(), $event->getExpungeDeletes());
@@ -269,7 +269,7 @@ class BufferedAdd extends AbstractPlugin
         $this->clear();
 
         $event = new PostCommitEvent($result);
-        $this->client->getEventDispatcher()->dispatch(Events::POST_COMMIT, $event);
+        $this->client->getEventDispatcher()->dispatch($event, Events::POST_COMMIT);
 
         return $result;
     }