+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the solarium library will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [5.1.2]
+### Fixed
+- BufferedAdd does not support Symfony event dispatcher
+
 ## [5.1.1]
 ### Fixed
 - PHP 7.1 compatibility issue: date constants are not available as part of DateTimeInterface before PHP 7.2.0